rename annotation info link to annotation object id link
