--- frontend/packages/data-portal/app/components/Run/AnnotationObjectTable/components/InfoLink/index.ts
@@ -1,1 +1,0 @@
-export { InfoLink } from './InfoLink'
