+++ frontend/packages/data-portal/app/components/Run/AnnotationObjectTable/components/ObjectIdLink/index.ts
@@ -1,0 +1,1 @@
+export { ObjectIdLink } from './ObjectIdLink'
